pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "connection_hub"
description = "Manages lobbies and player connections in the games."
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "uuid7==0.1.*",
    "redis==5.2.*",
    "nats-py==2.9.*",
    "httpx==0.28.*",
    "dishka==1.4.*",
    "adaptix==3.0.b9",
    "cyclopts==3.0.*",
    "fastapi[standard]==0.115.*",
    "uvicorn[standard]==0.34.*",
    "gunicorn==23.0.*",
]

[project.optional-dependencies]
dev = [
    "mypy==1.14.*",
    "ruff==0.8.*",
    "pre-commit==4.0.*",
    "pytest==8.3.*",
    "pytest-asyncio==0.25.*",
]

[project.scripts]
connection-hub = "connection_hub.main.cli:main"
